+++ src/loginServices.js
@@ -4,6 +4,7 @@ import rtlDetect from 'rtl-detect';
 import moment from 'moment';
 
 import { discoverServices } from './discoverServices';
+import { resetStore } from './mainActions';
 
 import {
   clearCurrentUser,
@@ -24,6 +25,7 @@ import {
   updateCurrentUser,
 } from './okapiActions';
 import processBadResponse from './processBadResponse';
+import StripesSession from './StripesSession';
 
 // export supported locales, i.e. the languages we provide translations for
 export const supportedLocales = [
@@ -56,6 +58,8 @@ export const supportedLocales = [
   'zh-TW',  // chinese, taiwan
 ];
 
+const SESSION_NAME = 'okapiSess';
+
 // export supported numbering systems, i.e. the systems tenants may chose
 // for numeral display
 export const supportedNumberingSystems = [
@@ -69,10 +73,9 @@ export const userLocaleConfig = {
   'module': '@folio/stripes-core',
 };
 
-function getHeaders(tenant, token) {
+function getHeaders(tenant) {
   return {
     'X-Okapi-Tenant': tenant,
-    'X-Okapi-Token': token,
     'Content-Type': 'application/json',
   };
 }
@@ -164,8 +167,11 @@ export function loadTranslations(store, locale, defaultTranslations = {}) {
  * @returns {Promise}
  */
 function dispatchLocale(url, store, tenant) {
-  return fetch(url,
-    { headers: getHeaders(tenant, store.getState().okapi.token) })
+  return fetch(url, {
+    headers: getHeaders(tenant),
+    credentials: 'include',
+    mode: 'cors',
+  })
     .then((response) => {
       if (response.status === 200) {
         response.json().then((json) => {
@@ -240,8 +246,11 @@ export function getUserLocale(okapiUrl, store, tenant, userId) {
  * @returns {Promise}
  */
 export function getPlugins(okapiUrl, store, tenant) {
-  return fetch(`${okapiUrl}/configurations/entries?query=(module==PLUGINS)`,
-    { headers: getHeaders(tenant, store.getState().okapi.token) })
+  return fetch(`${okapiUrl}/configurations/entries?query=(module==PLUGINS)`, {
+    headers: getHeaders(tenant),
+    credentials: 'include',
+    mode: 'cors',
+  })
     .then((response) => {
       if (response.status < 400) {
         response.json().then((json) => {
@@ -266,8 +275,11 @@ export function getPlugins(okapiUrl, store, tenant) {
  * @returns {Promise}
  */
 export function getBindings(okapiUrl, store, tenant) {
-  return fetch(`${okapiUrl}/configurations/entries?query=(module==ORG and configName==bindings)`,
-    { headers: getHeaders(tenant, store.getState().okapi.token) })
+  return fetch(`${okapiUrl}/configurations/entries?query=(module==ORG and configName==bindings)`, {
+    headers: getHeaders(tenant),
+    credentials: 'include',
+    mode: 'cors',
+  })
     .then((response) => {
       let bindings = {};
       if (response.status >= 400) {
@@ -348,12 +360,26 @@ export function spreadUserWithPerms(userWithPerms) {
 }
 
 /**
+ * logout
+ * dispatch events to clear the store, then clear the session too.
+ *
+ * @param {object} redux store
+ *
+ * @returns {Promise}
+ */
+export async function logout(store) {
+  store.dispatch(clearOkapiToken());
+  store.dispatch(clearCurrentUser());
+  store.dispatch(resetStore());
+  return localforage.removeItem(SESSION_NAME);
+}
+
+/**
  * createOkapiSession
  * Remap the given data into a session object shaped like:
  * {
  *   user: { id, username, personal }
  *   perms: { permNameA: true, permNameB: true, ... }
- *   token: token
  * }
  * Dispatch the session object, then return a Promise that fetches
  * and dispatches tenant resources.
@@ -361,12 +387,11 @@ export function spreadUserWithPerms(userWithPerms) {
  * @param {*} okapiUrl
  * @param {*} store
  * @param {*} tenant
- * @param {*} token
  * @param {*} data
  *
  * @returns {Promise}
  */
-export function createOkapiSession(okapiUrl, store, tenant, token, data) {
+export function createOkapiSession(okapiUrl, store, tenant, data) {
   // clear any auth-n errors
   store.dispatch(setAuthError(null));
 
@@ -380,53 +405,29 @@ export function createOkapiSession(okapiUrl, store, tenant, token, data) {
 
   const sessionTenant = data.tenant || tenant;
   const okapiSess = {
-    token,
+    // @@ still gotta provide `token` because RootWithIntl relies on it
+    token: '>>> MAKE COOKIES, NOT TOKENS <<<',
     user,
     perms,
     tenant: sessionTenant,
+    tokenExpiration: data.tokenExpiration,
   };
 
+  // configure a StripesSession object which will set up a timer internally
+  // and auto-logout unless it detects user interaction
+  //
+  // eslint-disable-next-line no-unused-vars
+  const ss = new StripesSession(store, data.tokenExpiration, logout);
+
   return localforage.setItem('loginResponse', data)
-    .then(() => localforage.setItem('okapiSess', okapiSess))
+    .then(() => localforage.setItem(SESSION_NAME, okapiSess))
     .then(() => {
       store.dispatch(setSessionData(okapiSess));
       return loadResources(okapiUrl, store, sessionTenant, user.id);
     });
 }
 
-/**
- * validateUser
- * return a promise that fetches from bl-users/self.
- * if successful, dispatch the result to create a session
- * if not, clear the session and token.
- *
- * @param {string} okapiUrl
- * @param {redux store} store
- * @param {string} tenant
- * @param {object} session
- *
- * @returns {Promise}
- */
-export function validateUser(okapiUrl, store, tenant, session) {
-  const { token, user, perms, tenant: sessionTenant = tenant } = session;
 
-  return fetch(`${okapiUrl}/bl-users/_self`, { headers: getHeaders(sessionTenant, token) }).then((resp) => {
-    if (resp.ok) {
-      return resp.json().then((data) => {
-        store.dispatch(setLoginData(data));
-        store.dispatch(setSessionData({ token, user, perms, tenant: sessionTenant }));
-        return loadResources(okapiUrl, store, sessionTenant, user.id);
-      });
-    } else {
-      store.dispatch(clearCurrentUser());
-      store.dispatch(clearOkapiToken());
-      return localforage.removeItem('okapiSess');
-    }
-  }).catch((error) => {
-    store.dispatch(setServerDown());
-    return error;
-  });
-}
 
 /**
  * getSSOEnabled
@@ -438,7 +439,9 @@ export function validateUser(okapiUrl, store, tenant, session) {
  * @returns {Promise}
  */
 export function getSSOEnabled(okapiUrl, store, tenant) {
-  return fetch(`${okapiUrl}/saml/check`, { headers: { 'X-Okapi-Tenant': tenant, 'Accept': 'application/json' } })
+  return fetch(`${okapiUrl}/saml/check`, {
+    headers: { 'X-Okapi-Tenant': tenant, 'Accept': 'application/json' }
+  })
     .then((response) => {
       if (response.status >= 400) {
         store.dispatch(checkSSO(false));
@@ -502,7 +505,7 @@ function processSSOLoginResponse(resp) {
  * @returns {Promise} resolving to the response's JSON
  */
 export function handleLoginError(dispatch, resp) {
-  return localforage.removeItem('okapiSess')
+  return localforage.removeItem(SESSION_NAME)
     .then(() => processBadResponse(dispatch, resp))
     .then(responseBody => {
       dispatch(setOkapiReady());
@@ -518,18 +521,16 @@ export function handleLoginError(dispatch, resp) {
  * @param {redux store} store
  * @param {string} tenant
  * @param {Response} resp HTTP response
- * @param {string} ssoToken
  *
  * @returns {Promise} resolving with login response body, rejecting with, ummmmm
  */
-export function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
-  const token = resp.headers.get('X-Okapi-Token') || ssoToken;
+export function processOkapiSession(okapiUrl, store, tenant, resp) {
   const { dispatch } = store;
 
   if (resp.ok) {
     return resp.json()
       .then(json => {
-        return createOkapiSession(okapiUrl, store, tenant, token, json)
+        return createOkapiSession(okapiUrl, store, tenant, json)
           .then(() => json);
       })
       .then((json) => {
@@ -539,6 +540,53 @@ export function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
   } else {
     return handleLoginError(dispatch, resp);
   }
+}
+
+/**
+ * validateUser
+ * return a promise that fetches from bl-users/self.
+ * if successful, dispatch the result to create a session
+ * if not, clear the session and token.
+ *
+ * @param {string} okapiUrl
+ * @param {redux store} store
+ * @param {string} tenant
+ * @param {object} session
+ *
+ * @returns {Promise}
+ */
+export function validateUser(okapiUrl, store, tenant, session) {
+  const { user, perms, tenant: sessionTenant = tenant, token, tokenExpiration } = session;
+
+  return fetch(`${okapiUrl}/bl-users/_self`, {
+    headers: getHeaders(sessionTenant),
+    credentials: 'include',
+    mode: 'cors',
+  }).then((resp) => {
+    if (resp.ok) {
+      console.log('>>> validateUser::resp.ok');
+      return resp.json().then((data) => {
+        console.log('session', session);
+
+        // configure a StripesSession object which will set up a timer internally
+        // and auto-logout unless it detects user interaction
+        //
+        // eslint-disable-next-line no-unused-vars
+        const ss = new StripesSession(store, session.tokenExpiration, logout);
+
+        store.dispatch(setLoginData(data));
+        store.dispatch(setSessionData({ user, perms, tenant: sessionTenant, token, tokenExpiration }));
+        return loadResources(okapiUrl, store, sessionTenant, user.id);
+      });
+    } else {
+      console.error('>>> validateUser !resp.ok');
+      return logout(store);
+    }
+  }).catch((error) => {
+    console.error('validateUser', error);
+    store.dispatch(setServerDown());
+    return error;
+  });
 }
 
 /**
@@ -552,7 +600,7 @@ export function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
  * @param {string} tenant
  */
 export function checkOkapiSession(okapiUrl, store, tenant) {
-  localforage.getItem('okapiSess')
+  localforage.getItem(SESSION_NAME)
     .then((sess) => {
       return sess !== null ? validateUser(okapiUrl, store, tenant, sess) : null;
     })
@@ -576,10 +624,12 @@ export function checkOkapiSession(okapiUrl, store, tenant) {
  * @returns {Promise}
  */
 export function requestLogin(okapiUrl, store, tenant, data) {
-  return fetch(`${okapiUrl}/bl-users/login?expandPermissions=true&fullPermissions=true`, {
-    method: 'POST',
-    headers: { 'X-Okapi-Tenant': tenant, 'Content-Type': 'application/json' },
+  return fetch(`${okapiUrl}/bl-users/login-with-expiry?expandPermissions=true&fullPermissions=true`, {
     body: JSON.stringify(data),
+    credentials: 'include',
+    headers: { 'X-Okapi-Tenant': tenant, 'Content-Type': 'application/json' },
+    method: 'POST',
+    mode: 'cors',
   })
     .then(resp => processOkapiSession(okapiUrl, store, tenant, resp));
 }
@@ -589,15 +639,15 @@ export function requestLogin(okapiUrl, store, tenant, data) {
  * retrieve currently-authenticated user
  * @param {string} okapiUrl
  * @param {string} tenant
- * @param {string} token
  *
  * @returns {Promise} Promise resolving to the response of the request
  */
-function fetchUserWithPerms(okapiUrl, tenant, token) {
-  return fetch(
-    `${okapiUrl}/bl-users/_self?expandPermissions=true&fullPermissions=true`,
-    { headers: getHeaders(tenant, token) },
-  );
+function fetchUserWithPerms(okapiUrl, tenant) {
+  return fetch(`${okapiUrl}/bl-users/_self?expandPermissions=true&fullPermissions=true`, {
+    headers: getHeaders(tenant),
+    credentials: 'include',
+    mode: 'cors',
+  });
 }
 
 /**
@@ -606,13 +656,12 @@ function fetchUserWithPerms(okapiUrl, tenant, token) {
  * @param {string} okapiUrl
  * @param {redux store} store
  * @param {string} tenant
- * @param {string} token
  *
  * @returns {Promise} Promise resolving to the response-body (JSON) of the request
  */
-export function requestUserWithPerms(okapiUrl, store, tenant, token) {
-  return fetchUserWithPerms(okapiUrl, tenant, token)
-    .then(resp => processOkapiSession(okapiUrl, store, tenant, resp, token));
+export function requestUserWithPerms(okapiUrl, store, tenant) {
+  return fetchUserWithPerms(okapiUrl, tenant)
+    .then(resp => processOkapiSession(okapiUrl, store, tenant, resp));
 }
 
 /**
@@ -648,10 +697,10 @@ export function requestSSOLogin(okapiUrl, tenant) {
  * @returns {Promise}
  */
 export function updateUser(store, data) {
-  return localforage.getItem('okapiSess')
+  return localforage.getItem(SESSION_NAME)
     .then((sess) => {
       sess.user = { ...sess.user, ...data };
-      return localforage.setItem('okapiSess', sess);
+      return localforage.setItem(SESSION_NAME, sess);
     })
     .then(() => {
       store.dispatch(updateCurrentUser(data));
@@ -669,8 +718,9 @@ export function updateUser(store, data) {
  */
 export async function updateTenant(okapi, tenant) {
   const okapiSess = await localforage.getItem('okapiSess');
-  const userWithPermsResponse = await fetchUserWithPerms(okapi.url, tenant, okapi.token);
+  const userWithPermsResponse = await fetchUserWithPerms(okapi.url, tenant);
   const userWithPerms = await userWithPermsResponse.json();
 
-  await localforage.setItem('okapiSess', { ...okapiSess, tenant, ...spreadUserWithPerms(userWithPerms) });
+  await localforage.setItem(SESSION_NAME, { ...okapiSess, tenant, ...spreadUserWithPerms(userWithPerms) });
 }
+