+++ src/useOkapiKy.js
@@ -2,18 +2,19 @@ import ky from 'ky';
 import { useStripes } from './StripesContext';
 
 export default () => {
-  const { locale = 'en', tenant, token, url } = useStripes().okapi;
+  const { locale = 'en', tenant, url } = useStripes().okapi;
   return ky.create({
-    prefixUrl: url,
+    credentials: 'include',
     hooks: {
       beforeRequest: [
         request => {
           request.headers.set('Accept-Language', locale);
           request.headers.set('X-Okapi-Tenant', tenant);
-          request.headers.set('X-Okapi-Token', token);
         }
       ]
     },
+    mode: 'cors',
+    prefixUrl: url,
     retry: 0,
     timeout: 30000,
   });